documentation of feature_wise_hessian and removed bug in hessian with size factors
before, size factors was implicitly broadcasted to the dimension of the data matrix X, this is a relatively new feature. however, this had not been adapted in the feature wise slicing in feature_wise_hessian where still the full size factor matrix was fed into the computation of the hessian of a single gene (slice of X). Accordingly, the other matrices were automatically broadcasted to the full size of X during the computation of the hessian of each X which led to extremely high memory load during hessian estimation. These high memory loads do not occur anymore now.